added SetElevatorCommand, changed constants
